apps/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Joyce Server"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Punit Arani", email = "punitsai36@gmail.com"}
]

dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "chromadb>=1.0.20",
    "fastapi[standard]>=0.116.1",
    "livekit-agents[assemblyai,cartesia,google,openai,silero,turn-detector]",
    "livekit-plugins-noise-cancellation",
    "openai>=1.102.0",
    "pydantic>=2.11.7",
    "python-dotenv",
    "sqlalchemy[asyncio]>=2.0.43",
    "supabase>=2.18.1",
    "tabulate>=0.9.0",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "python-dotenv>=1.1.1",
    "ruff",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "DTZ", "T10", "DJ", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["S101", "T201", "PLR2004", "UP006", "UP045", "TRY003", "EM101", "ANN001", "ANN201", "ANN202", "F841", "COM812", "SLF001", "ARG001", "PERF401", "S311", "B007", "PT018", "E501", "PIE808", "F541", "RUF003", "TC001", "TC002", "TC003"]

[tool.ruff.lint.pylint]
max-args = 8

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
